VB, TA, RJ, AM:
We have taken the unusual (one-time) step to modify the structure
of the ELPA git repository as follows:

1) The current ELPA version has been designated as a "stable"
   version. It has proven to be rock solid for us so far, and
   deserves this name.

2) We have decided on a stable version naming scheme,
   "ELPA_year.month". The first such version is
   ELPA_2011.12. This version will only ever be updated
   for clear and unambiguous bug fixes.

3) All further development will happen in the directory named
   ELPA_development_version. At this time, ELPA_2011.12 is an
   exact copy of ELPA_development_version.

No substantial changes have been made to either repository as yet.
(But we are happy to say that the present "ELPA" publication, which
we hope ELPA users will cite, now has its final volume and page
numbers.)

We apologize if this change causes any inconvenience for you. Please
try to make sure that any changes you might make really affect the
development version, and not the stable version. We do not plan any
structural changes of a similar scope at any point in the future,
but it had to be done once.

Happy ELPA stable version day!