BUGFIX and BUGFIX versioning
A bug in ELPA_2013.08 was fixed: due to typo in a preprocessor statement
the kernels optimized for BLUGENE/P could not be used correctly.

In course of this action, also a bugfix versioning scheme was
introduced: in case of bugfixes they are versioned with an extra 3 digit
number in the ELPA release name, i.e ELPA 2013.08 -> 2013.08.001

Consistently any new release of ELPA will from now on start on the minor
level "000".