Debug output of test cases
For debugging purposes the test examples of the ELPA library can
now write the Eigenvectors in an ascii file, e.g
a call of "test_real 4000 1500 16 output" will envoke the test
case for real matrices of the ELPA1 library. A matrix of 4000x4000
values will be created an 1500 EV's will be computet with a 16 blocking.
And the results will be --- due to the output flag --- written to a
text file