ELPA_2013.11 release (minor version 000)
This is the release of the ELPA_development_version_OpenMP
If OpenMP support is not used, this version has the same
functionality as ELPA_2013.08.

If OpenMP support is used, obviously, a hybrid version of ELPA will be
build.

Allthough this is a release, version ELPA_2013.11 is far from complete!
During the next week optimizations of the OpenMP part will be published,
however, the basic functionality is set by this commit