ELPA_development_version_qr: Changes in OpenMP detection
The Intel Fortran compiler accepts the flag "-fopenmp" for compilation
with OpenMP. However, the Intel MPI compiler wrapper does not.
With the Intel compiler, this leads to the fact, that if ELPA is compiled
with the "-fopenmp" flag a not thread-save version of the Intel MPI
library is used and the test (with make check) fails.

Intel promised to solve this in the future.

However, for now the problem is solved in the user friendly way that no
manipulation of the MPI compiler wrappers have to be done:

For detecting the OpenMP compiler flags, instead of the predefined
macro "AC_OPENMP" of autoconf a modified macro "AX_ELPA_OPENMP"
is used, which first checks "-openmp" and only then "-fopenmp".
Thus it is ensured that the Intel compiler (and mpi compiler wrapper)
does not get confused.

This is invisible for users calling "configure" during the installation
process.