Another ABI change in ELPA_2014.06 pre-release
Build servers recommend against calling "abort" functions in a
library. Thus, from now on, ELPA  returns a success flag, which
is a change of ABI!

In order to force the user to honour this, the ELPA drivers
are converted to functions, which implies an adaptation and
rebuild of the user code. The driver functions return a
logical "success", which, if it is false, indicates an
error within the library.

It is advised to check catch and check this
return value in the user code, which calls ELPA.