Fix libtool library version and naming
It was decided that the library file shall be called

  libelpa.so.*

without an explicit version string. Version information will be encoded
by using the -version-info parameter of libtool, resulting in an actual
filename as defined by your flavor of Unix.

The file LIBRARY_INTERFACE documents the changes, the current interface
has number 1 and is binary incompatible with the previous version 0.