Final(?) version number scheme change
It was decided that the version number shall be all of 2014.06.001, that
is including the release version.

Sadly, pkg-config is not too flexible with different versions of
libraries, therefore the pkg-config file will have the full version
number including the release. This might be a bit cumbersome for
dependent projects.