configure.ac: treat GPU kernel as other kernels
Configure treats the GPU kernels now as any other kernel, i. e.
if GPU support is enabled (and it is possible to build it) then
it will be build in ADDITION to all other possible kernels for
the desired hardware.

Also, it is possbile to configure the build process for
the GPU version ONLY (as it was already possible to trigger the
build for only ONE specific real/complex kernel).

Note: The sources at the moment CANNOT handle this, i.e. if
GPU support is configured, the GPU only code path is compiled.
This will be changed in the near future.