Split file elpa2.F90 into elpa2.F90 and elpa2_compute.F90
As in a previous commit for elpa1.F90, for automatic generation of
documentation elpa2.F90 has been splitted in two files, in order to
have a lean, easy-to-understand user interface:

elpa2.F90
the visible user functions, which provide the library calls.
The usage is the same as before

elpa2_compute.F90
all internal routines, which are used by ELPA2, but which are never
called external of the library by a user. These functions are now
"hidden" in the module elpa2_compute, which is used by ELPA2.

The procedures in elpa2_compute.F90 are identical to the ones in
elpa2.F90 before this split commit. The only -- but quite a lot of them
-- changes are intendation changes.