Fix actual reason for linker problems
The actual reason for the linker problems was that the Fortran libraries
where listed before the object files, by modifying the _LINK automake
variables. The proper way to do is of course to add the necessary
libraries after the object files by appending them to the _LDADD
variables.

As the MPI module was not responsible for the linker problems it is now
used by default, unless explicitly switched off with

  --disable-mpi-module

Also, the C test programs that had these linker errors where previously
not compiled in the OpenMP case, for no obvious reason. Now they are
also included there.