Better control of AVX kernels
Sometimes it might be desirable that only one of the specifc AVX
kernels (AVX, AVX2, AVX512) are build, even though the processor
supports all (or several) of these instructions.

The flag --disable-[avx | avx2 | avx512] allows to define which kernels
should NOT be build. The default is: what can be build will be build