Workaround for PGI compiler
PGI's compiler wants

  -module

as the flag to specify the module output directory. Sadly, this is also
a libtool argument, resulting in much sorrow.

Thus, we have to escape -module by prefixing it with -Xcompiler. Sadly
libtool is not used everywhere by automake, thus we have to remove it
again in those cases were it would be passed to the compiler and cause
sorrow there.

Sad story. But with a happy ending?