Prevent compiler error due to uninitialized value
Another blocker for the SuSE build server: Even though the variable is
later not actually used by the function, a recent GCC complains about
the use of an uninitialized value for that variable.

This prevents build on the OBS, as they use compiliation flags that
detect such behaviour.

Error message there:

 ../test/C/driver/legacy_interface/legacy_real_driver_c_version.c:193:12: warning: 'THIS_REAL_ELPA_KERNEL_API' is used uninitialized in this function [-Wuninitialized]
     success = elpa_solve_evp_real_double(na, nev, a, na_rows, ev, z, na_rows, nblk, na_cols, mpi_comm_rows, mpi_comm_cols, my_mpi_comm_world, THIS_REAL_ELPA_KERNEL_API, useQr, useGPU,

 [...]

 I: Program is using uninitialized variables.
    Note the difference between "is used" and "may be used"
 W: elpa uninitialized-variable ../test/C/driver/legacy_interface/legacy_real_driver_c_version.c:193