Had to move some functions to the public namespace
The PGI compiler (of course) complained about a missing module
(elpa_generated_fortran_interfaces.mod) when compiling the test
programs.

It is true (in a way) that some part of this module is indeed necessary,
as the public-facing function signatures have arguments that are
referring to those three functions in their type:

  elpa_strerr_c(elpa_error)
  elpa_int_value_to_string_c(name, value, string)
  elpa_int_value_to_strlen_c(name, value)

Thus, for these three we create another header prefix,

  !pf>

for Fortran definitions that should be public. Those are included in
elpa_api.F90.