automated processing of scaling results
I also included text file, which is the product of this processing for
the scaling performed this summer and can be used to retrieve all the
data needed for plots. It also contains references to original
log-files, so it is easy to find where is each number going from. Of
course, one needs the original raw data (log files), which are at the
moment only on my hydra account. Maybe we should put them somewhere more
visible?