tile size can be influence by the user
tile size is used in elpa1_tridiag and elpa2_bandred.
Iits value is computed based on block size and least common multiple of
number of processes rows and colums, but then can be enlarged to some
multiple of this value. How much it should be enlarged can now be
influenced by min_tile_size parameter.