altering the hh_transform_real to fix the issue from the previous commit
HOWEVER, this is probably not a final solution, since it might break
some other calculations (matrices). It is a bit dangerous.

The problem arises by the way elpa ignores small numbers in this
routine, see the comment:

      ! Similar to LAPACK routine DLARFP, but uses ||x||**2 instead of x(:)
      ! and returns the factor xf by which x has to be scaled.
      ! It also hasn't the special handling for numbers < 1.d-300 or > 1.d150
      ! since this would be expensive for the parallel implementation.

when the originally used (and immediately suspicious) condition
     if ( XNORM_SQ==0.0_rk ) then
is used, one can see from the introduced print, that indeed very small
numbers (XNORM_SQ < 1e-250_rk) appear after some time. If instead
      if ( XNORM_SQ < 1e-100_rk ) then
is used, this does not happen. But, it is not clear how to choose the
threshold and what else can be affected.

The lapack routine DLARFP uses a more sophisticated handling of small
numbers, but it is not applicable in elpa, since we only have XNORM_SQ
in the routine and not the whole vector X as in lapack.