Fix error in elpa_index
lfind() expects a proper comparison function where the order of the
arguments does not matter. Previously it was assumed that the first
argument is always a pointer to an unspecific "key" object and the
second argument a pointer to an array member to compare against.

However, the standard requires the two arguments to be interchangeable
pointers to array-member like objects.

Surfaced as musl-libc does in fact call it the other way around,
compared to glibc.