adding test which splits communicator at the beginning
and runs several elpas (still mpi-based) in parallel. Each elpa
"instance" then splits its communicator to row and column communicator,
as it is always done. This is actually often done in FHI-aims and was
previously leading to issues when using OmniPath. Hopefully was solved
by Intel MPI 2019.3. Keeping the test for future reference.