HETEROGENOUS_CLUSTER support
On heterogenous cluster, of nodes with different CPUs the _experimental_
feature (--enable-heterogenous-cluster-support) can be used:

It compares the (Intel) cpuid set of all CPUs which are used by ELPA MPI
processes and finds the SIMD instruction set, which is supported by all
used CPUs. The ELPA 2stage back-transformation kernel (a.k.a "kernel")
will be set accordingly on all MPI processes.

This feature, can override the setting of the kernel done previously by
the user!

At the moment it will only work for Intel CPUs, i.e. clusters consisting
of nodes with Intel CPUs and e.g. AMD CPUs are at the moment _NOT_
supported.

Since this is an experimental feature, it might be dropped again in the
future, if it turns out not to be useful for the users