Fix the MPI communicators per elpa object
Clarification of the ELPA usage:
It has always been intended with the ELPA API that one should
only set the MPI communicators ("mpi_comm_parent", "mpi_comm_rows",
and "mpi_comm_cols") _once_ per ELPA object.
Technically, it has been possible to change these communicators, for
an existing ELPA object, which leads -- dependent on the exact
configuration of the ELPA object -- to correct or erroneous behaviour.

With this commit, it is _technically_ impossible to set the
communicators more than once (per ELPA object)