configure.ac: Use a function prototype for _ linkage probing
The prototype chosen (char return type, no arguments) follows the
usual autoconf approach for link tests.  This is why compilers
are likely to continue to support this construct even if there is
a mismatch in types.

Future compilers are likely to reject implicit function declarations,
and the checks will always fail with such compilers.  (Implicit
function declarations were removed from C in 1999.)